python/candidate-shortlist/requirements.txt
certifi==2025.11.12
fastapi==0.121.1
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
SQLAlchemy==2.0.44
uvicorn==0.38.0
psycopg2-binary==2.9.11